Overwrite original lib.rs with migrated lib2.rs.
